Tree files renamed to address that they cover binary tree, or binary search tree algorithms
