+++ algorithms/trees/README.md
@@ -40,8 +40,8 @@ A trie is a tree-like data structure used to store strings, where each node repr
 For Tree operations we will use Classes and Functions available in the [helper.py](/algorithms/trees/helper.py) script.
 
 Currently covered Tree operations are:
-* Traversals: [traversals.py](/algorithms/trees/traversals.py)
-* Insert into BST: [insert_bst.py](/algorithms/trees/insert_bst.py)
+* Traversals: [traversals.py](/algorithms/trees/bt_traversals.py)
+* Insert into BST: [insert_bst.py](/algorithms/trees/bst_insert.py)
 * Search BST: [search_bst.py](/algorithms/search/search_bst.py)
-* Delete Node from BST: [delete_node_bst.py](/algorithms/trees/delete_node_bst.py)
+* Delete Node from BST: [delete_node_bst.py](/algorithms/trees/bst_delete_node.py)
 * Calculate Tree height: [bt_height.py](/algorithms/trees/bt_height.py)
