+++ algorithms/trees/bst_insert.py
@@ -1,5 +1,5 @@
 from helper import TreeNode
-from traversals import in_order_traversal
+from bt_traversals import in_order_traversal
 
 
 def insert_into_bst(root, value):
